+++ R/isosurface.R
@@ -150,6 +150,71 @@ algebraicMesh <- function(
   cgalMesh$new(clean = xptr)
 }
 
+#' @title Extract mesh from a NifTI file.
+#' @description Make a triangle mesh from a NifTI file and an isovalue.
+#'
+#' @param fileName path to the NifTI file (extension \code{nii} 
+#'   or \code{nii.gz})
+#' @param isolevel the level at which to construct the isosurface
+#' @param sphereCenter,sphereRadius center and radius of a bounding sphere
+#' @param angleBound lower bound in degrees for the angles of the faces of 
+#'   the mesh
+#' @param radiusBound upper bound for the radii of surface Delaunay balls; 
+#'   a surface Delaunay ball is a ball circumscribing a face and centered 
+#'  at the surface
+#' @param distanceBound upper bound for the distance between the 
+#'   circumcenter of a face and the center of the surface Delaunay 
+#'   ball of this face
+#'
+#' @return A \code{cgalMesh} object.
+#' @export
+#' @importFrom RNifti readNifti writeAnalyze
+#' @examples 
+#' library(cgalMeshes)
+#' library(rgl)
+#' teapot <- system.file("extdata", "teapot.nii.gz", package = "cgalMeshes")
+#' mesh <- voxel2mesh(
+#'   teapot, isolevel = 6,
+#'   sphereCenter = c(128, 128, 89), 
+#'   sphereRadius = 200,
+#'   angleBound    = 20, 
+#'   radiusBound   = 5, 
+#'   distanceBound = 5
+#' )
+#' mesh$computeNormals()
+#' rmesh <- rotate3d(mesh$getMesh(), pi, 1, 0, 0)
+#' open3d(windowRect = 50 + c(0, 0, 512, 512))
+#' view3d(0, 0)
+#' shade3d(rmesh, color = "red")
+#' wire3d(rmesh)
+voxel2mesh <- function(
+    fileName, isolevel,
+    sphereCenter, sphereRadius,
+    angleBound, radiusBound, distanceBound
+) {
+  stopifnot(isFilename(fileName))
+  stopifnot(isNumber(isolevel))
+  stopifnot(isVector3(sphereCenter))
+  stopifnot(isPositiveNumber(sphereRadius))
+  stopifnot(isPositiveNumber(angleBound))
+  stopifnot(isPositiveNumber(radiusBound))
+  stopifnot(isPositiveNumber(distanceBound))
+  checkFile <- grepl("\\.nii\\.gz$", tolower(fileName)) || 
+    grepl("\\.nii$", tolower(fileName))
+  if(!checkFile) {
+    stop("Not a NifTI file.")
+  }
+  hdrFile <- tempfile(fileext = ".hdr")
+  img <- readNifti(fileName)
+  . <- writeAnalyze(img, hdrFile)
+  xptr <- VoxelToMesh(
+    hdrFile, isolevel, 
+    sphereCenter, 
+    sphereRadius,
+    angleBound, radiusBound, distanceBound
+  )
+  cgalMesh$new(clean = xptr)
+}
 
 #' Title
 #'
@@ -191,7 +256,7 @@ isocontour <- function(
   writeVolume(bvol, niiFile)
   img <- readNifti(niiFile)
   . <- writeAnalyze(img, hdrFile)
-  xptr <- Isomesh(
+  xptr <- VoxelToMesh(
     hdrFile, isolevel, 
     sphereCenter - c(x[1L], y[1L], z[1L]), 
     sphereRadius,